[Refactor] JPQL 에 의한 총 좋아요 수 형 변환

+++ backend/src/main/java/com/synergy/backend/domain/likes/model/response/LikesInfoResponse.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 @Builder
 public class LikesInfoResponse {
     private boolean memberIsLiked;
-    private int productLikesCount;
+    private Long productLikesCount;
     private int atelierHavingProductsLikesCount;
 }
